parallel-benchmark: Split into smaller files
